requirements.txt
cycler==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.0
pandas==1.2.5
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.0
scitools-pyke==1.1.1
seaborn==0.11.1
six==1.16.0
switch==1.1.0
threadpoolctl==2.1.0
